requirements.txt
pycoingecko>=3.1.0,<4.0
pandas>=2.1.0,<3.0
numpy>=1.26,<2.2
pyyaml>=6.0,<7.0
scikit-learn>=1.4,<2.0
joblib>=1.3,<2.0
mlflow>=2.13,<3.0
streamlit>=1.35,<2.0
plotly>=5.20,<6.0
dvc>=3.50,<4.0
ray[tune]>=2.10,<3.0
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
